Eliminate separate issuer and verifier classes
Gafaelfawr had generic JWT issuer and verifier classes, left over
from when it used JWT both internally and externally and therefore
had generic JWT code.  Move the remaining code into the OIDC
provider (for handling tokens upstream of us) and the OIDC service
(for issuing and verifying our own tokens), since there was almost
no common code between the two.

For now, keep the configuration structure the same and pass the
higher-level config object into each class.
