+++ tests/providers/oidc_test.py
@@ -1,4 +1,4 @@
-"""Tests for the gafaelfawr.verify package."""
+"""Tests for verification of upstream OpenID Connect token verification."""
 
 from __future__ import annotations
 
@@ -23,10 +23,11 @@ from gafaelfawr.exceptions import (
 from gafaelfawr.factory import ComponentFactory
 from gafaelfawr.keypair import RSAKeyPair
 from gafaelfawr.models.oidc import OIDCToken
+from gafaelfawr.providers.oidc import OIDCProvider
 
-from .support.oidc import mock_oidc_provider_config
-from .support.settings import configure
-from .support.tokens import create_upstream_oidc_token
+from ..support.oidc import mock_oidc_provider_config
+from ..support.settings import configure
+from ..support.tokens import create_upstream_oidc_token
 
 
 def encode_token(
@@ -46,12 +47,13 @@ def encode_token(
 
 
 @pytest.mark.asyncio
-async def test_verify_oidc(
+async def test_verify_token(
     tmp_path: Path, respx_mock: respx.Router, factory: ComponentFactory
 ) -> None:
     config = await configure(tmp_path, "oidc")
     factory.reconfigure(config)
-    verifier = factory.create_token_verifier()
+    provider = factory.create_provider()
+    assert isinstance(provider, OIDCProvider)
 
     now = datetime.now(timezone.utc)
     exp = now + timedelta(days=24)
@@ -66,27 +68,27 @@ async def test_verify_oidc(
 
     # Missing iss.
     with pytest.raises(InvalidIssuerError) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     assert str(excinfo.value) == "No iss claim in token"
 
     # Missing kid.
     payload["iss"] = "https://bogus.example.com/"
     token = encode_token(payload, keypair)
     with pytest.raises(UnknownKeyIdException) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     assert str(excinfo.value) == "No kid in token header"
 
     # Unknown issuer.
     token = encode_token(payload, keypair, kid="a-kid")
     with pytest.raises(InvalidIssuerError) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     assert str(excinfo.value) == "Unknown issuer: https://bogus.example.com/"
 
     # Unknown kid.
     payload["iss"] = config.verifier.oidc_iss
     token = encode_token(payload, keypair, kid="a-kid")
     with pytest.raises(UnknownKeyIdException) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     expected = f"kid a-kid not allowed for {config.verifier.oidc_iss}"
     assert str(excinfo.value) == expected
 
@@ -95,7 +97,7 @@ async def test_verify_oidc(
     kid = config.verifier.oidc_kids[0]
     token = encode_token(payload, config.issuer.keypair, kid=kid)
     with pytest.raises(MissingClaimsException) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     expected = f"No {config.verifier.username_claim} claim in token"
     assert str(excinfo.value) == expected
 
@@ -105,9 +107,9 @@ async def test_verify_oidc(
     await mock_oidc_provider_config(respx_mock, keypair)
     payload[config.verifier.username_claim] = "some-user"
     token = encode_token(payload, config.issuer.keypair, kid=kid)
-    verified_token = await verifier.verify_oidc_token(token)
+    verified_token = await provider.verify_token(token)
     with pytest.raises(MissingClaimsException) as excinfo:
-        verifier.get_uid_from_token(verified_token)
+        provider.get_uid_from_token(verified_token)
     expected = f"No {config.verifier.uid_claim} claim in token"
     assert str(excinfo.value) == expected
 
@@ -119,7 +121,8 @@ async def test_verify_oidc_no_kids(
     config = await configure(tmp_path, "oidc-no-kids")
     factory.reconfigure(config)
     keypair = config.issuer.keypair
-    verifier = factory.create_token_verifier()
+    provider = factory.create_provider()
+    assert isinstance(provider, OIDCProvider)
     await mock_oidc_provider_config(respx_mock, keypair, "kid")
 
     now = datetime.now(timezone.utc)
@@ -132,7 +135,7 @@ async def test_verify_oidc_no_kids(
     }
     token = encode_token(payload, keypair, kid="a-kid")
     with pytest.raises(UnknownKeyIdException) as excinfo:
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
     expected = f"Issuer {config.verifier.oidc_iss} has no kid a-kid"
     assert str(excinfo.value) == expected
 
@@ -144,7 +147,8 @@ async def test_key_retrieval(
     config = await configure(tmp_path, "oidc-no-kids")
     factory.reconfigure(config)
     assert config.oidc
-    verifier = factory.create_token_verifier()
+    provider = factory.create_provider()
+    assert isinstance(provider, OIDCProvider)
 
     # Initial working JWKS configuration.
     jwks = config.issuer.keypair.public_key_as_jwks("some-kid")
@@ -158,13 +162,13 @@ async def test_key_retrieval(
 
     # Check token verification with this configuration.
     token = await create_upstream_oidc_token(kid="some-kid")
-    assert await verifier.verify_oidc_token(token)
+    assert await provider.verify_token(token)
 
     # Wrong algorithm for the key.
     jwks.keys[0].alg = "ES256"
     respx_mock.get(jwks_url).respond(json=jwks.dict())
     with pytest.raises(UnknownAlgorithmException):
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
 
     # Should go back to working if we fix the algorithm and add more keys.
     # Add an explicit 404 from the OpenID connect endpoint.
@@ -173,18 +177,18 @@ async def test_key_retrieval(
     keypair = RSAKeyPair.generate()
     jwks.keys.insert(0, keypair.public_key_as_jwks("a-kid").keys[0])
     respx_mock.get(jwks_url).respond(json=jwks.dict())
-    assert await verifier.verify_oidc_token(token)
+    assert await provider.verify_token(token)
 
     # Try with a new key ID and return a malformed reponse.
     respx_mock.get(jwks_url).respond(json=["foo"])
     token = await create_upstream_oidc_token(kid="malformed")
     with pytest.raises(FetchKeysException):
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
 
     # Return a 404 error.
     respx_mock.get(jwks_url).respond(404)
     with pytest.raises(FetchKeysException):
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
 
     # Fix the JWKS handler but register a malformed URL as the OpenID Connect
     # configuration endpoint, which should be checked first.
@@ -193,8 +197,8 @@ async def test_key_retrieval(
     respx_mock.get(oidc_url).respond(json=["foo"])
     token = await create_upstream_oidc_token(kid="another-kid")
     with pytest.raises(FetchKeysException):
-        await verifier.verify_oidc_token(token)
+        await provider.verify_token(token)
 
     # Try again with a working OpenID Connect configuration.
     respx_mock.get(oidc_url).respond(json={"jwks_uri": jwks_url})
-    assert await verifier.verify_oidc_token(token)
+    assert await provider.verify_token(token)
